Important fix for 64-bit systems
Prevents segfault when loading a save file.

There are likely more bugs lurking in 64-bit land (such as one involving
memory allocation issues/leakage when loading full highscore files), but
at least this makes the game playable in 64-bit environments.

+++ Main/Source/char.cpp
@@ -1892,7 +1892,7 @@ void character::Load(inputfile& SaveFile)
     SaveFile >> TemporaryStateCounter[c];
 
   if(SaveFile.Get())
-    SetTeam(game::GetTeam(ReadType<int>(SaveFile)));
+    SetTeam(game::GetTeam(ReadType<ulong>(SaveFile)));
 
   if(SaveFile.Get())
     GetTeam()->SetLeader(this);
